Remove AppDomain isolation on more compiler tests
