+++ src/Compilers/VisualBasic/Test/Syntax/app.config
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <appSettings>
-    <add key="xunit.appDomain" value="required"/>
+    <add key="xunit.appDomain" value="denied"/>
     <add key="xunit.diagnosticMessages" value="false"/>
     <add key="xunit.parallelizeTestCollections" value="false"/>
   </appSettings>
